refactor(stage-tamagotchi): merge steam client types and read app id from env
